+++ CHANGELOG.md
@@ -5,6 +5,17 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 [Keep a CHANGELOG](http://keepachangelog.com).
 
 
+## Unreleased
+
+### Fixed
+
+- The `:update_request_method` was not correctly fetched from the application config.
+
+
+[Compare v0.4.1...HEAD](https://github.com/rdf-elixir/sparql_client/compare/v0.4.1...HEAD)
+
+
+
 ## v0.4.1 - 2022-11-03
 
 This version is just upgraded to RDF.ex 1.0.